+++ src/components/computation/shaders/index.ts
@@ -1,10 +1,14 @@
+import applyColorScaleChunk from '@/components/textures/shaders/applyColorScale.glsl';
 import MaxFrag from './Max.glsl'
 import MinFrag from './Min.glsl'
 import MeanFrag from './Mean.glsl'
 import StDevFrag from './StDev.glsl'
 import vertShader from './vert.glsl'
-import fragShader from './frag.glsl'
+import fragShaderRaw from './frag.glsl'
 import correlateFrag from './Correlation.glsl'
+
+const fragShader = fragShaderRaw.replace('// APPLY_COLOR_SCALE', applyColorScaleChunk);
+
 export {
     MaxFrag,
     MinFrag,
